Replace inconsistent use of index and id throughout.
Also fix types (use std::size_t)

+++ modules/phase_field/src/postprocessors/GrainTrackerInterface.C
@@ -18,7 +18,7 @@ InputParameters validParams<GrainTrackerInterface>()
   params.addParam<bool>("compute_op_maps", false, "Indicates whether the data structures that"
                                                   "hold the active order parameter information"
                                                   "should be populated or not");
-  params.addParam<unsigned int>("reserve_op", 0, "Indicates the number of reserved ops (variables that cannot be remapped to)");
+  params.addParam<unsigned short>("reserve_op", 0, "Indicates the number of reserved ops (variables that cannot be remapped to)");
   params.addParam<Real>("reserve_op_threshold", 0.95, "Threshold for locating a new feature on the reserved op variable(s)" );
   params.addParam<UserObjectName>("ebsd_reader", "Optional: EBSD Reader for initial condition");
   params.addParam<unsigned int>("phase", "EBSD phase number from which to retrieve information");
